// src/Types/upload.ts
export type Extensions = | 'jpg'
    | 'jpeg'
    | 'png'
    | 'webp'
    | 'gif'
    | 'mp3'
    | 'ogg'
    | 'wav'
    | 'm4a'
    | 'mp4'
    | 'mkv'
    | 'webm'
    | 'pdf'
    | 'zip'
    | 'json'

export type MimeTypes = | 'image/jpeg'
    | 'image/jpeg'
    | 'image/png'
    | 'image/webp'
    | 'image/gif'
    | 'audio/mpeg'
    | 'audio/ogg'
    | 'audio/wav'
    | 'audio/mp4'
    | 'video/mp4'
    | 'video/x-matroska'
    | 'video/webm'
    | 'application/pdf'
    | 'application/zip'
    | 'application/json';

export type MediaTypesStringExample = | 'audio.'
    | 'image.'
    | 'default.'
    | 'video.'
    | string;